+++ config/jest/test.config.js
@@ -2,12 +2,28 @@
  * This is the main jest config.  It runs tests using the default
  * test environment: jest-environment-jsdom.
  */
-const path = require("path");
+const fs = require("node:fs");
+const path = require("node:path");
+const ancesdir = require("ancesdir");
+
+const staticRootDir = ancesdir(__dirname);
+
+const swcrc = JSON.parse(
+    fs.readFileSync(path.join(staticRootDir, ".swcrc"), "utf8"),
+);
+// NOTE: We need to use this plugin in order to turn the module exports into
+// module.exports. This will make it so that we can mock exports correctly.
+// Check it out - Nullish coalescing _assignment_!
+// https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Operators/Nullish_coalescing_assignment
+swcrc.jsc ??= {};
+swcrc.jsc.experimental ??= {};
+swcrc.jsc.experimental.plugins ??= [];
+swcrc.jsc.experimental.plugins.push(["swc_mut_cjs_exports", {}]);
 
 module.exports = {
     rootDir: path.join(__dirname, "../../"),
     transform: {
-        "^.+\\.(j|t)sx?$": "<rootDir>/config/jest/test.transform.js",
+        "^.+\\.(t|j)sx?$": ["@swc/jest", swcrc],
         // Compile .svg files using a custom transformer that returns the
         // basename of the file being transformed.
         "^.+.svg$": "<rootDir>/config/jest/svg.transform.js",