+++ packages/wonder-blocks-button/src/themes/default.ts
@@ -5,14 +5,10 @@ const {semanticColor} = tokens;
 // The underline-offset is the distance between the text baseline and the
 // bottom of the underline. This is necessary to prevent the underline from
 // breaking with descenders.
-const textUnderlineOffset = tokens.spacing.xxxSmall_4;
+const textUnderlineOffset = tokens.sizing.size_040;
 
 const focusOutline = {
     border: semanticColor.focus.outer,
-};
-
-const focusOutlineLight = {
-    border: semanticColor.border.inverse,
 };
 
 const theme = {
@@ -25,83 +21,14 @@ const theme = {
             progressive: {
                 ...semanticColor.action.primary.progressive,
                 focus: focusOutline,
-                disabled: {
-                    background:
-                        semanticColor.action.primary.disabled.background,
-                    foreground:
-                        semanticColor.action.primary.disabled.foreground,
-                },
             },
 
-            // kind=primary / color=default / light=true
-            // NOTE: These colors will be removed from WB as soon as we remove the
-            // light variant.
-            progressiveLight: {
-                ...semanticColor.action.secondary.progressive,
-                default: {
-                    ...semanticColor.action.secondary.progressive.default,
-                    background: semanticColor.surface.primary,
-                },
-                focus: focusOutlineLight,
-                hover: {
-                    ...semanticColor.action.secondary.progressive.hover,
-                    border: semanticColor.border.inverse,
-                    backgrond: semanticColor.surface.primary,
-                },
-                press: {
-                    ...semanticColor.action.secondary.progressive.press,
-                    border: semanticColor.action.secondary.progressive.press
-                        .background,
-                },
-                disabled: {
-                    background:
-                        semanticColor.action.secondary.progressive.press
-                            .background,
-                    foreground:
-                        semanticColor.action.secondary.progressive.default
-                            .foreground,
-                },
-            },
             // kind=primary / color=destructive / light=false
             destructive: {
                 ...semanticColor.action.primary.destructive,
                 focus: focusOutline,
-                disabled: {
-                    background:
-                        semanticColor.action.primary.disabled.background,
-                    foreground:
-                        semanticColor.action.primary.disabled.foreground,
-                },
             },
-            // kind=primary / color=destructive / light=true
-            // NOTE: These colors will be removed from WB as soon as we remove the
-            // light variant.
-            destructiveLight: {
-                ...semanticColor.action.secondary.destructive,
-                default: {
-                    ...semanticColor.action.secondary.destructive.default,
-                    background: semanticColor.surface.primary,
-                },
-                focus: focusOutlineLight,
-                hover: {
-                    ...semanticColor.action.secondary.progressive.hover,
-                    border: semanticColor.border.inverse,
-                    background: semanticColor.surface.primary,
-                },
-                press: {
-                    ...semanticColor.action.secondary.destructive.press,
-                    border: semanticColor.action.secondary.destructive.press
-                        .background,
-                },
-                disabled: {
-                    background:
-                        semanticColor.action.secondary.destructive.press
-                            .background,
-                    foreground:
-                        semanticColor.action.secondary.destructive.default
-                            .foreground,
-                },
-            },
+            disabled: semanticColor.action.primary.disabled,
         },
 
         /**
@@ -119,49 +46,8 @@ const theme = {
                     // NOTE: This is a special case for the secondary button
                     icon: "transparent",
                 },
-                disabled: {
-                    // NOTE: This is a special case for the secondary button
-                    // TODO(WB-1895): Revisit disabled styles.
-                    border: semanticColor.action.primary.disabled.border,
-                    foreground:
-                        semanticColor.action.secondary.disabled.foreground,
-                },
             },
-            // kind=secondary / color=default / light=true
-            // NOTE: These colors will be removed from WB as soon as we remove the
-            // light variant.
-            progressiveLight: {
-                default: {
-                    border: tokens.color.white64,
-                    background: "transparent",
-                    foreground: semanticColor.text.inverse,
-                },
-                focus: focusOutlineLight,
-                hover: {
-                    border: semanticColor.border.inverse,
-                    background: "transparent",
-                    foreground: semanticColor.text.inverse,
-                    // NOTE: Not used, but included for type safety.
-                    icon: "transparent",
-                },
-                press: {
-                    border: tokens.color.fadedBlue,
-                    background:
-                        semanticColor.action.primary.progressive.press
-                            .background,
-                    foreground: semanticColor.text.inverse,
-                },
-                disabled: {
-                    border: semanticColor.action.secondary.progressive.press
-                        .background,
-                    background:
-                        semanticColor.action.secondary.progressive.press
-                            .background,
-                    // NOTE: Using primitive token, but this will go away once
-                    // we remove the light variant.
-                    foreground: tokens.color.white50,
-                },
-            },
+
             // kind=secondary / color=destructive / light=false
             destructive: {
                 ...semanticColor.action.secondary.destructive,
@@ -171,46 +57,13 @@ const theme = {
                     // NOTE: This is a special case for the secondary button
                     icon: "transparent",
                 },
-                disabled: {
-                    // NOTE: This is a special case for the secondary button
-                    // TODO(WB-1895): Revisit disabled styles.
-                    border: semanticColor.action.primary.disabled.border,
-                    foreground:
-                        semanticColor.action.secondary.disabled.foreground,
-                },
             },
-            // kind=secondary / color=destructive / light=true
-            // NOTE: These colors will be removed from WB as soon as we remove the
-            // light variant.
-            destructiveLight: {
-                default: {
-                    border: tokens.color.white64,
-                    background: "transparent",
-                    foreground: semanticColor.text.inverse,
-                },
-                focus: focusOutlineLight,
-                hover: {
-                    border: semanticColor.border.inverse,
-                    background: "transparent",
-                    foreground: semanticColor.text.inverse,
-                    // NOTE: Not used, but included for type safety.
-                    icon: "transparent",
-                },
-                press: {
-                    border: tokens.color.fadedRed,
-                    background:
-                        semanticColor.action.primary.destructive.press
-                            .background,
-                    foreground: semanticColor.text.inverse,
-                },
-                disabled: {
-                    border: semanticColor.action.secondary.destructive.press
-                        .background,
-                    background:
-                        semanticColor.action.secondary.destructive.press
-                            .background,
-                    foreground: tokens.color.white50,
-                },
+
+            disabled: {
+                ...semanticColor.action.secondary.disabled,
+                // NOTE: This is a special case for the secondary button
+                // TODO(WB-1895): Revisit disabled styles.
+                border: semanticColor.action.primary.disabled.border,
             },
         },
         /**
@@ -223,66 +76,15 @@ const theme = {
             progressive: {
                 ...semanticColor.action.tertiary.progressive,
                 focus: focusOutline,
-                disabled: {
-                    foreground:
-                        semanticColor.action.tertiary.disabled.foreground,
-                },
             },
-            // kind=tertiary / color=default / light=true
-            // NOTE: These colors will be removed from WB as soon as we remove the
-            // light variant.
-            progressiveLight: {
-                default: {
-                    border: tokens.color.white64,
-                    background: "transparent",
-                    foreground: semanticColor.text.inverse,
-                },
-                focus: focusOutlineLight,
-                hover: {
-                    border: semanticColor.border.inverse,
-                    background: "transparent",
-                    foreground: semanticColor.text.inverse,
-                },
-                press: {
-                    border: semanticColor.border.inverse,
-                    foreground: tokens.color.fadedBlue,
-                },
-                disabled: {
-                    foreground: tokens.color.white50,
-                },
-            },
+
             // kind=tertiary / color=destructive / light=false
             destructive: {
                 ...semanticColor.action.tertiary.destructive,
                 focus: focusOutline,
-                disabled: {
-                    foreground:
-                        semanticColor.action.tertiary.disabled.foreground,
-                },
             },
-            // kind=tertiary / color=destructive / light=true
-            // NOTE: These colors will be removed from WB as soon as we remove the
-            // light variant.
-            destructiveLight: {
-                default: {
-                    border: tokens.color.white64,
-                    background: "transparent",
-                    foreground: semanticColor.text.inverse,
-                },
-                focus: focusOutlineLight,
-                hover: {
-                    border: semanticColor.border.inverse,
-                    background: "transparent",
-                    foreground: semanticColor.text.inverse,
-                },
-                press: {
-                    border: semanticColor.border.inverse,
-                    foreground: tokens.color.fadedRed,
-                },
-                disabled: {
-                    foreground: tokens.color.white50,
-                },
-            },
+
+            disabled: semanticColor.action.tertiary.disabled,
         },
     },
     border: {
@@ -339,12 +141,13 @@ const theme = {
     font: {
         size: {
             // NOTE: This token is specific to this button size.
-            large: 18,
+            large: "1.8rem",
         },
         lineHeight: {
-            small: tokens.font.lineHeight.small + textUnderlineOffset,
-            default: tokens.font.lineHeight.medium + textUnderlineOffset,
-            large: tokens.font.lineHeight.medium + 2 + textUnderlineOffset,
+            small: tokens.font.lineHeight.xMedium,
+            default: tokens.font.lineHeight.large,
+            // NOTE: this token is specific to this button size.
+            large: "2.6rem",
         },
         weight: {
             default: tokens.font.weight.bold,