+++ packages/wonder-blocks-clickable/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @khanacademy/wonder-blocks-clickable
 
+## 7.0.3
+
+### Patch Changes
+
+- Updated dependencies [2656fd4]
+- Updated dependencies [6018552]
+- Updated dependencies [7bbf311]
+- Updated dependencies [7f79943]
+    - @khanacademy/wonder-blocks-tokens@9.0.0
+
 ## 7.0.2
 
 ### Patch Changes